background_tasks/requirements.txt
apscheduler==3.6.3
blinker==1.4
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.1.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
PyMySQL==0.10.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
stripe==2.55.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3